Move logic in request route to utilities functions
